Update database schema and ORM settings
The existing SQL init script for Postgres has been extended to include more tables, constraints, and indexes that mirror the application's intended usage and operations. Also, the Hibernate ORM configuration in application.properties was changed from 'update' to 'validate' signalling that the application is now in a mature stage where database schema changes are to be handled manually.